feat(projects-gallery): add initial project request
